refactor: divide types
