+++ src/types/profile.tsx
@@ -1,0 +1,6 @@
+import type { ConvertForm } from './store';
+
+export interface ProfileProps {
+  data: ConvertForm;
+  last?: boolean;
+}
